+++ packages/adp-tooling/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/adp-tooling
 
+## 0.9.27
+
+### Patch Changes
+
+-   Updated dependencies [312919ec]
+    -   @sap-ux/axios-extension@1.12.0
+    -   @sap-ux/system-access@0.3.24
+
 ## 0.9.26
 
 ### Patch Changes